+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - car.json by default avoids private roads
 - maxspeed<5 is ignored, maxspeed=none is ignored with some exceptions, maxspeed parsing and related constants were renamed #3077
 - Added agency_id, agency_name, and route_type to Trip.PtLeg to describe the operator and the vehicle type used in public transport.
+- Fixed distance value for "pt" legs in the response of the /route endpoint.
+- Added shape geometry for "pt" legs in the response of the /route endpoint.
 
 ### 10.0 [5 Nov 2024]